run_stubtest.py
"""Run stubtest with Django configured."""

import os
import sys

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "tests.mypy_test_settings")

import django

django.setup()

from mypy.stubtest import main

sys.exit(main())
